Don't allow `@apply`ing classes that ever appear inside of an at-rule
This is too complex to easily support; better to explicitly error for
now vs. the current behavior which is just silently doing something
other than you probably expect.
